fix file upload bugs

+++ controllers/users_controller.js
@@ -24,7 +24,10 @@ module.exports.update = async function (req, res) {
         user.email = req.body.email;
         if (req.file) {
           if (user.avatar) {
-            fs.unlinkSync(path.join(__dirname, "..", user.avatar));
+            if (fs.existsSync(path.join(__dirname, "..", user.avatar))) {
+              //deleting the file (old avatar)
+              fs.unlinkSync(path.join(__dirname, "..", user.avatar));
+            }
           }
 
           user.avatar = User.avatarPath + "/" + req.file.filename;
